jsonrpc: move the server code to a `server` module
And document the modules a bit!
